+++ internal/xerrors/stacktrace_test.go
@@ -1,8 +1,13 @@
 package xerrors
 
 import (
+	"errors"
 	"fmt"
 	"testing"
+
+	"github.com/stretchr/testify/require"
+	grpcCodes "google.golang.org/grpc/codes"
+	grpcStatus "google.golang.org/grpc/status"
 )
 
 func TestStackTraceError(t *testing.T) {
@@ -11,25 +16,71 @@ func TestStackTraceError(t *testing.T) {
 		text  string
 	}{
 		{
-			error: WithStackTrace(fmt.Errorf("TestError")),
+			error: WithStackTrace(fmt.Errorf("fmt.Errorf")),
 			//nolint:lll
-			text: "TestError at `github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors.TestStackTraceError(stacktrace_test.go:14)`",
+			text: "fmt.Errorf at `github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors.TestStackTraceError(stacktrace_test.go:19)`",
 		},
 		{
-			error: WithStackTrace(fmt.Errorf("TestError%s", "Printf")),
+			error: WithStackTrace(fmt.Errorf("fmt.Errorf %s", "Printf")),
 			//nolint:lll
-			text: "TestErrorPrintf at `github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors.TestStackTraceError(stacktrace_test.go:19)`",
+			text: "fmt.Errorf Printf at `github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors.TestStackTraceError(stacktrace_test.go:24)`",
 		},
 		{
-			error: WithStackTrace(fmt.Errorf("TestError%s", "Printf")),
+			error: WithStackTrace(
+				WithStackTrace(errors.New("errors.New")),
+			),
 			//nolint:lll
-			text: "TestErrorPrintf at `github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors.TestStackTraceError(stacktrace_test.go:24)`",
+			text: "errors.New at `github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors.TestStackTraceError(stacktrace_test.go:30)` at `github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors.TestStackTraceError(stacktrace_test.go:29)`",
 		},
 	} {
 		t.Run(test.text, func(t *testing.T) {
-			if test.error.Error() != test.text {
-				t.Fatalf("unexpected text of error: \"%s\", exp: \"%s\"", test.error.Error(), test.text)
-			}
+			require.Equal(t, test.text, test.error.Error())
+		})
+	}
+}
+
+func TestTransportStackTraceError(t *testing.T) {
+	for _, test := range []struct {
+		error error
+		text  string
+	}{
+		{
+			error: WithStackTrace(
+				Transport(
+					grpcStatus.Error(grpcCodes.Aborted, "some error"),
+					WithAddress("example.com"),
+				),
+			),
+			//nolint:lll
+			text: "transport/Aborted (\"rpc error: code = Aborted desc = some error\", address: \"example.com\") at `github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors.TestTransportStackTraceError(stacktrace_test.go:48)`",
+		},
+		{
+			error: WithStackTrace(
+				WithStackTrace(
+					Transport(
+						grpcStatus.Error(grpcCodes.Aborted, "some error"),
+					),
+				),
+			),
+			//nolint:lll
+			text: "transport/Aborted (\"rpc error: code = Aborted desc = some error\") at `github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors.TestTransportStackTraceError(stacktrace_test.go:59)` at `github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors.TestTransportStackTraceError(stacktrace_test.go:58)`",
+		},
+		{
+			error: WithStackTrace(
+				WithStackTrace(
+					WithStackTrace(
+						Transport(
+							grpcStatus.Error(grpcCodes.Aborted, "some error"),
+						),
+					),
+				),
+			),
+			//nolint:lll
+			text: "transport/Aborted (\"rpc error: code = Aborted desc = some error\") at `github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors.TestTransportStackTraceError(stacktrace_test.go:71)` at `github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors.TestTransportStackTraceError(stacktrace_test.go:70)` at `github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors.TestTransportStackTraceError(stacktrace_test.go:69)`",
+		},
+	} {
+		t.Run(test.text, func(t *testing.T) {
+			require.Equal(t, test.text, test.error.Error())
 		})
 	}
 }