+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * Added checks for nil option to all opts range loops
 * Moved content of package `internal/ctxlabels` into `internal/xcontext`
 * Implemented `GRPCStatus` method in `internal/xerrors/transportError`
+* Added different implementations of stacktrace error for grpc errors and other 
 
 ## v3.42.5
 * Fixed closing of `database/sql` connection (aka `YDB` session)